fix: resolve 9 full-matrix vector test failures via FP accumulation order fixes
Three floating-point evaluation order mismatches vs upstream C:

1. find_best_pitch (pitch.rs): Syy update used two separate operations
   `Syy += A; Syy -= B` instead of C's single `Syy += A - B`. Due to FP
   non-associativity, this caused pitch_search to return different results
   after ~864 calls, diverging the bitstream at packet 870.

2. gemm_accum (layers.rs): accumulated into a temporary before adding to
   out[i], whereas C accumulates directly into out[i].

3. leak_boost (analysis.rs): C's `.5` literal is a double, causing implicit
   f64 promotion. Rust was computing entirely in f32.

All 1460 full-matrix vector tests now pass (was 1451/1460).

+++ src/opus/analysis.rs
@@ -1370,10 +1370,10 @@ fn tonality_analysis(
                 band_log2[b as usize] - (leakage_from[b as usize] + 2.5f32)
             });
         tonal.info[info_idx].leak_boost[b as usize] =
-            (if (255) < (0.5 + (64.0 * boost)).floor() as i32 {
+            (if (255) < (0.5f64 + (64.0f32 * boost) as f64).floor() as i32 {
                 255
             } else {
-                (0.5 + (64.0 * boost)).floor() as i32
+                (0.5f64 + (64.0f32 * boost) as f64).floor() as i32
             }) as u8;
         b += 1;
     }
